CORDA-756 - Refactor Kryo into sub module as per AMQP
